minor changes to achieve backwards compatibility
git-svn-id: file:///srv/local/data/cleanRepo/trunk@11896 a0517364-e873-4167-b1c8-416daa5a6b91

+++ playgrounds/mrieser/src/main/java/playground/mrieser/core/sim/network/queueNetwork/QueueLink.java
@@ -134,7 +134,7 @@ import playground.mrieser.core.sim.network.api.SimLink;
 		if (position == SimLink.POSITION_AT_FROM_NODE) {
 			// vehicle enters from intersection
 			this.vehQueue.add(vehicle);
-			double earliestLeaveTime = now + this.freespeedTravelTime;
+			double earliestLeaveTime = (int) (now + this.freespeedTravelTime); // (int) for backwards compatibility
 			this.earliestLeaveTimes.put(vehicle, earliestLeaveTime);
 			this.usedStorageCapacity += vehicle.getSizeInEquivalents();
 			this.network.simEngine.getEventsManager().processEvent(
